+++ src/public-api.ts
@@ -180,7 +180,7 @@ const Gyos = {
 	portalCreate,
 	portalDestroy,
 	setCspNonce,
-	version: '0.3.0'
+	version: '0.3.1'
 };
 
 export default Gyos;